pyproject.toml
[project]
name = "network-security"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "certifi>=2025.8.3",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pymongo[srv]>=4.14.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
]
